+++ ricaun.RevitTest.Application/Revit/Utils/RibbonUtils.cs
@@ -10,6 +10,10 @@ namespace ricaun.RevitTest.Application.Revit
         public static string TestWait { get; } = GetComponentImage("test-wait");
         public static string GetComponentImage(string imageName)
         {
+            return string.Format("Images/{0}.png", imageName);
+        }
+        static string GetComponentImagePack(string imageName)
+        {
             var assemblyName = Assembly.GetCallingAssembly().GetName().Name;
             return string.Format("pack://application:,,,/{0};component/Images/{1}.png", assemblyName, imageName);
         }